core/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
requests==2.32.3
email-validator==2.2.0
